backend/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loan-extraction-backend"
version = "0.1.0"
description = "Loan document data extraction system backend"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "asyncpg>=0.31.0",
    "alembic>=1.18.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "structlog>=25.5.0",
    # Phase 2+ dependencies (include now to avoid conflicts later)
    "docling>=2.70.0",
    "google-genai>=1.60.0",
    "google-cloud-storage>=2.18.0",
    "tenacity>=9.0.0",
    # Phase 3 validation dependencies
    "rapidfuzz>=3.0.0",
    "phonenumbers>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.0.0",
    "mypy>=1.14.0",
    "ruff>=0.8.0",
    "aiosqlite>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false
disallow_incomplete_defs = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # Line length handled by formatter

[tool.ruff.format]
quote-style = "double"

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["*/__init__.py"]

[tool.coverage.report]
exclude_also = [
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
fail_under = 80
show_missing = true
